Move App to own component
